+++ packages/at_repl/CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 2.1.0
 - feat: `--key-file|-k option to specify key file`
 - feat: `--version` to print at_repl version
-- chore: Renamed `--rootUrl` to `--rootDomain` but kept `--rootUrl` as an
-  alias. `--root-domain` is also supported.
+- feat: root domain options: `--root-domain`, `--rootDomain`, `--root-url`, 
+  `--rootUrl`, `-r`
 - feat: support proxy servers in root domain
 
 ## 2.0.0